Added MiniDumpWriteDump and MINIDUMP_TYPE manually.
For some reason the ClangSharpPInvokeGenerator tool skips MiniDumpWriteDump and the MINIDUMP_TYPE enum so for now add them back manually until https://github.com/dotnet/ClangSharp/issues/571 is fixed.

Also this maps the last 3 parameters to the function to void* pointers to make the application developer responsible for providing the types to pass in to them if they want to provide them, otherwise they can pass null as usual.

Signed-off-by: AraHaan <seandhunt_7@yahoo.com>